neuroimage-ml/fsleyes/requirements.txt
numpy
nibabel
PyOpenGL
streamlit
nilearn